Eval: Load weights from directory instead of a certain checkpoint file
Allow the user to provide a path to the directory of all checkpoint
files instead of providing the explicit path of the checkpoint file.
An additional flag (TEST.USE_WEIGHTS_DIR) was added to the config file.
A utility function to perform the logic of checking this flag and
looking for the latest file in the specified WEIGHT_DIR was added also.

Signed-off-by: ep <ep@no.mail>