train output folder name: use "auto-sorted" datetime format
Change the directory name used for storing trained checkpoints. Lets say
we started a training on 28 October 2019, at 15:04, instead of using a
the following format "28Oct19_1504" which is easier for humans to read
but is not natively sorted, use a format such as 20191028_1504 which is
a bit less understandable for humans but we can say it is "auto-sorted"
since lexicographically and chronographically sort of entries are
equivalent when we use this representation.

Signed-off-by: ep <ep@no.mail>