Bug fix in utils.draw_bbox() function prototype
Before this commit, the utils.draw_bbox() had an argument `classes` with
a default value defined as 'classes=read_class_names(cfg.YOLO.CLASSES)'.
This means that the call to read_class_names() occurred on the import of
the utils module. This behaviour is error prone especially since the
read_class_names() may raise an exception dependant of the path set in
cfg.YOLO.CLASSES, which in case of relative path it depends also in the
current working directory that set during the import time.
The resolution was to change the classes default value to an empty list
and change the caller of utils.draw_bbox() to evaluate the result of
read_class_names(cfg.YOLO.CLASSES) and provide it explicitly.

Signed-off-by: ep <ep@no.mail>