Add a test to test reproducibility of the train procedure
The test uses a predefined seed vale to seed the random number
generators and initializes an empty network, and train/test it for a
reduced predefined number of epochs on a narrow set of images.
After the short train the test compare the train and test average step
loss against predefined values.
Another algorithmic fitting is to use a momentum optimizer instead of
Adam-optimizer which contains built in randomness which we did not find
a way to seed it (i.e. make it reproducible).

Signed-off-by: ep <ep@no.mail>